test: add tests for createEVMContext
